Working on the overall architecture
I'm creating a bundle of utility hooks for use by my projects.  I'm
starting out with a post-commit hook that will restore arbitrary
permissions on files that have been updated by git.

This is initially meant for my puppet configuration scheme, but I can
see other uses for it in the future.

The hooks that get run will be configurable on a per project basis.